Project Description
Project description
Awesome chocolates is a company that sell different kind of chocolates is different countries . And my task is to analyze that which chocolate category is selling least in which region by which team. And Which Chocolates category create the huge revenue in which reigon by which team. So that they can took the decision according to the given analysis.

Tool Used :-
1-Microsoft power bi
2-Microsoft excel

Techniques:-
1-Data cleaning
2-Building Data Model
3-Usind DAX and power pivot to measures and calculations
4-Creating and changing visuals
5-Data visualization